requirement.txt
locust==2.37.4
pyyaml>=6.0
influxdb==5.3.2
